Improve double hashing.
Because of the double_hash == 0 check, the step size of 1 would have
twice the frequency of any other.  With the table size being prime
numbers, any integer is co-prime with the table size, so any nonzero
step number will end up hitting every element of the hash table.

Based on a patch by Andrea Canciani to the X Server.