Standardize language for license blurbs.
This extends the language from the source files into the Makefiles as well.

The most significant change is from "ADAM JACKSON" to "THE AUTHORS OR
COPYRIGHT HOLDERS". While I'm sure that Adam would appreciate any
reduction in liability, the intent of this language is to disclaim
liability for the actual authors and copyright holders.

(I don't know that Adam actually has any copyright interest in this
implementation, but if he does, he's still covered by the wider
language.)